prefer TMPDIR over /tmp as default socket path for compatibility with secure platforms with file segregation  and no shared /tmp path
this is equivalent to

--socket-path=$TMPDIR/.virgl_test

 # back ground

it's a security hole if apps can share data . the android emulator has /data/local/tmp but not release builds. terminals can only write in their own folder. /tmp does not exist in any android terminal without a cumbersome ptrace hack to change / which is really never necessary